# api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NotNetflix",
    "description": "This is the backend for the NotNetflix application.",
    "version": "0.3.4",
    "licence": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "http://www.localhost:4000",
  "basePath": "/",
...
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/roles": {
      "post": {
        "tags": ["Role"],
        "summary": "Create role",
        "description": "Create role",
        "operationId": "CreateRole",
        "parameters": [
          {
            "name": "name",
            "in": "body",
            "description": "Name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful create role",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "date-time"
                },
                "updatedAt": {
                  "type": "date-time"
                }
              }
            }
          },
          "400": {
            "description": "Invalid role name"
          }
        }
      },
      "get":{
        "tags": ["Roles"],
        "summary": "Get roles",
        "description": "Get roles",
        "operationId": "GetRole",
        "responses": {
          "200": {
            "description": "Successful get role",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "date-time"
                },
                "updatedAt": {
                  "type": "date-time"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": ["User"],
        "summary": "Get users",
        "description": "Get users",
        "operationId": "GetUsers",
        "responses": {
          "200": {
            "description": "Successful get users",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "Cannot get users"
          }
        }
      },
      "post": {
        "tags": ["User"],
        "summary": "Create user",
        "description": "Create user",
        "operationId": "CreateUser",
        "parameters": [
          {
            "name": "name",
            "in": "body",
            "description": "Name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "body",
            "description": "Email",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "body",
            "description": "Password",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful create user",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "date-time"
                },
                "updatedAt": {
                  "type": "date-time"
                }
              }
            }
          },
          "400": {
            "description": "Invalid user name or email or password"
          }
        }
      }
    },
    "/users/{id}":{
      "get": {
        "tags": ["User"],
        "summary": "Get user",
        "description": "Get user by id",
        "operationId": "GetUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful get user",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                },
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "date-time"
                },
                "updatedAt": {
                  "type": "date-time"
                },
                "roleId": {
                  "type": "integer"
                }
              }
            }
          },
          "400": {
            "description": "Resource not found"
          }
        }
      },
      "patch": {
        "tags": ["User"],
        "summary": "Update user",
        "description": "Update user",
        "operationId": "UpdateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "email",
            "in": "body",
            "description": "Email",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "body",
            "description": "Password",
            "required": true,
            "schema": {
              "type": "string"
            }
          }]
      }
    },
...
  }
}